pyproject.toml
# Copyright 2025 Canonical Ltd.
# See LICENSE file for licensing details.
[project]
name = "owasp-logger"
version = "0.1"
requires-python = "~=3.8.0"

dependencies = [
  "pydantic",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-instrumentation",
]

[dependency-groups]
dev = [
  # Linting
  "ruff",
  "codespell",
  # Static
  "pyright",
  # Unit
  "pytest",
  "coverage[toml]",
]

# Testing tools configuration
[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

# Formatting tools configuration
[tool.black]
line-length = 99
target-version = ["py38"]

# Linting tools configuration
[tool.ruff]
line-length = 99
extend-exclude = ["__pycache__", "*.egg_info"]

[lint.pydocstyle]
convention = "google"

[tool.pyright]
extraPaths = ["src", "examples"]
pythonVersion = "3.8"
pythonPlatform = "Linux"

[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
asyncio_mode = "auto"
addopts = "--tb=native --verbose --capture=no --log-cli-level=INFO"
markers = ["setup", "work", "teardown"]

[tool.codespell]
skip = ".git,.tox,build,venv*"
ignore-words-list = "assertIn"
